testing - init

+++ Modules/Adjustment/Database/Migrations/2021_07_22_004043_create_adjusted_products_table.php
@@ -13,17 +13,19 @@ class CreateAdjustedProductsTable extends Migration
      */
     public function up()
     {
-        Schema::create('adjusted_products', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('adjustment_id');
-            $table->unsignedBigInteger('product_id');
-            $table->unsignedBigInteger('id_location')->nullable();
-            $table->unsignedBigInteger('sub_location')->nullable();
-            $table->integer('quantity');
-            $table->string('type');
-            $table->foreign('adjustment_id')->references('id')->on('adjustments')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (app()->environment() !== 'testing') {
+            Schema::create('adjusted_products', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('adjustment_id');
+                $table->unsignedBigInteger('product_id');
+                $table->unsignedBigInteger('id_location')->nullable();
+                $table->unsignedBigInteger('sub_location')->nullable();
+                $table->integer('quantity');
+                $table->string('type');
+                $table->foreign('adjustment_id')->references('id')->on('adjustments')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
