+++ Modules/Product/Database/Migrations/2023_09_14_144054_add_new_foreign_produk_id_to_product_items_table.php
@@ -13,17 +13,19 @@ class AddNewForeignProdukIdToProductItemsTable extends Migration
      */
     public function up()
     {
-        Schema::table('product_items', function (Blueprint $table) {
-          DB::statement('SET FOREIGN_KEY_CHECKS=0;');
-
-          
-         $table->foreign('product_id')->references('id')
-                  ->on('products')->onDelete('cascade');
-
-
-            DB::statement('SET FOREIGN_KEY_CHECKS=0;');   
-
-        });
+        if (app()->environment() !== 'testing') {
+            Schema::table('product_items', function (Blueprint $table) {
+              DB::statement('SET FOREIGN_KEY_CHECKS=0;');
+    
+              
+             $table->foreign('product_id')->references('id')
+                      ->on('products')->onDelete('cascade');
+    
+    
+                DB::statement('SET FOREIGN_KEY_CHECKS=0;');   
+    
+            });
+        }
     }
 
     /**
